+++ apps/main/src/app/baseline/_api/index.ts
@@ -1,56 +1,17 @@
-import { db } from '@repo/database';
 import { cacheLife } from 'next/cache';
-import { getBlogMetadata } from '@/services/blogs/blog';
-import { getBlogs } from '@/services/blogs/blogs';
+import { getBaselineFeatures as _getBaselineFeatures } from '@/services/baseline/baseline';
+import { getFeatureBlogMap as _getFeatureBlogMap } from '@/services/baseline/feature-blog-map';
 
-export type BaselineFeature = {
-  featureId: string;
-  name: string;
-  status: 'newly' | 'widely';
-  date: string;
-};
-
-export type BlogLink = {
-  slug: string;
-  title: string;
-};
-
-export async function getBaselineFeatures(): Promise<{
-  features: BaselineFeature[];
-}> {
+export async function getBaselineFeatures() {
   'use cache';
   cacheLife('minutes');
 
-  const snapshots = await db.query.baselineSnapshots.findMany();
-
-  const features = snapshots
-    .map((s) => ({
-      featureId: s.featureId,
-      name: s.name,
-      status: s.status as 'newly' | 'widely',
-      date: s.date,
-    }))
-    .toSorted((a, b) => b.date.localeCompare(a.date));
-
-  return { features };
+  return await _getBaselineFeatures();
 }
 
-export async function getFeatureBlogMap(): Promise<Record<string, BlogLink>> {
+export async function getFeatureBlogMap() {
   'use cache';
   cacheLife('max');
 
-  const blogs = await getBlogs();
-  const map: Record<string, BlogLink> = {};
-
-  await Promise.all(
-    blogs.map(async (blog) => {
-      const metadata = await getBlogMetadata(blog.slug);
-      if (!metadata.featureIds) return;
-      for (const featureId of metadata.featureIds) {
-        map[featureId] = { slug: blog.slug, title: metadata.title };
-      }
-    }),
-  );
-
-  return map;
+  return await _getFeatureBlogMap();
 }
